+++ tutorials/nft/frontend-app/public/index.html
