+++ tutorials/nft/frontend-app/public/manifest.json
