+++ tutorials/nft/frontend-app/src/components/HomePage.js
