+++ tutorials/nft/frontend-app/src/components/dialogs/PurchaseNFTTokenDialog.js
