+++ tutorials/nft/frontend-app/src/components/dialogs/TransferFundsDialog.js
