+++ tutorials/nft/frontend-app/src/fixture.txt
