+++ tutorials/nft/frontend-app/src/utils/common.js
