+++ tutorials/nft/frontend-app/src/utils/transactions/create_nft_token.js
